# docs/swagger.yaml
basePath: /
definitions:
...
  ent.Task:
    properties:
      description:
        description: Description holds the value of the "description" field.
        type: string
      done:
        description: Done holds the value of the "done" field.
        type: boolean
      done_date_time:
        description: DoneDateTime holds the value of the "done_date_time" field.
        type: string
      due_date_time:
        description: DueDateTime holds the value of the "due_date_time" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      notes:
        description: Notes holds the value of the "notes" field.
        type: string
      tags:
        description: Tags holds the value of the "tags" field.
        items:
          type: string
        type: array
    type: object
...
  handlers.ChangeTasksOrderRequest:
    properties:
      containerID:
        type: integer
      containerType:
        $ref: '#/definitions/schema.ContainerType'
      tasksInNewOrder:
        items:
          type: integer
        type: array
    required:
    - containerID
    - containerType
    - tasksInNewOrder
    type: object
  handlers.IDsRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
...
  handlers.NewHierarchicalTasksRequest:
    properties:
      nodes:
        items:
          $ref: '#/definitions/models.HierarchicalTaskNode'
        type: array
      parent:
        $ref: '#/definitions/models.ContainerDescription'
    required:
    - nodes
    - parent
    type: object
...
  handlers.NewTaskRequest:
    properties:
      parent:
        $ref: '#/definitions/models.ContainerDescription'
      task:
        $ref: '#/definitions/models.TaskShort'
    type: object
...
  handlers.PatchTaskByIDRequest:
    properties:
      description:
        type: string
      notes:
        type: string
    type: object
...
  handlers.TaskIDRequest:
    properties:
      _id:
        type: integer
    type: object
  handlers.TaskResponse:
    properties:
      actions:
        items:
          type: integer
        type: array
      definitions:
        items:
          type: integer
        type: array
      description:
        type: string
      done:
        type: boolean
      done_date_time:
        type: string
      id:
        type: integer
      knowledge_bits:
        items:
          type: integer
        type: array
      knowledge_nodes:
        items:
          type: integer
        type: array
      notes:
        type: string
      parent_containers:
        items:
          items: {}
          type: array
        type: array
      problems:
        items:
          type: integer
        type: array
      questions:
        items:
          type: integer
        type: array
      tags:
        items:
          type: string
        type: array
    type: object
...
  models.ContainerCheck:
    properties:
      containerID:
        type: integer
      containerType:
        $ref: '#/definitions/schema.ContainerType'
      description:
        type: string
      id:
        type: integer
    type: object
  models.ContainerDescription:
    description: Container description with ID and type
    properties:
      id:
        example: 1
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/schema.ContainerType'
        example: task
    type: object
...
  models.HierarchicalTaskNode:
    properties:
      children:
        items:
          $ref: '#/definitions/models.HierarchicalTaskNode'
        type: array
      description:
        type: string
    type: object
...
  models.TaskFull:
    properties:
      actions:
        items:
          type: integer
        type: array
      checks:
        items:
          $ref: '#/definitions/models.ContainerCheck'
        type: array
      definitions:
        items:
          type: integer
        type: array
      description:
        type: string
      done:
        type: boolean
      doneDateTime:
        type: string
      dueDateTime:
        type: string
      id:
        type: integer
      knowledgeBits:
        items:
          type: integer
        type: array
      knowledgeNodes:
        items:
          type: integer
        type: array
      longTasks:
        items:
          type: integer
        type: array
      notes:
        type: string
      parentContainers:
        items:
          $ref: '#/definitions/models.ContainerDescription'
        type: array
      problems:
        items:
          type: integer
        type: array
      questions:
        items:
          type: integer
        type: array
      tags:
        items:
          type: string
        type: array
      tasks:
        items:
          type: integer
        type: array
      variables:
        items:
          $ref: '#/definitions/models.ContainerVariable'
        type: array
    type: object
  models.TaskPartial:
    properties:
      description:
        type: string
      done:
        type: boolean
      doneDateTime:
        type: string
      dueDateTime:
        type: string
      id:
        type: integer
      notes:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
  models.TaskReportTreeNode:
    properties:
      children:
        items:
          $ref: '#/definitions/models.TaskReportTreeNode'
        type: array
      depth:
        type: integer
      name:
        type: string
    type: object
...
  schema.ContainerType:
    enum:
    - epic
    - story
    - task
    - question
    - problem
    - knowledge-node
    - knowledge-bit
    - definition
    - action
    - repetitive-task
    - long-task
    - direction
    - state
    type: string
    x-enum-varnames:
    - ContainerTypeEpic
    - ContainerTypeStory
    - ContainerTypeTask
    - ContainerTypeQuestion
    - ContainerTypeProblem
    - ContainerTypeKnowledgeNode
    - ContainerTypeKnowledgeBit
    - ContainerTypeDefinition
    - ContainerTypeAction
    - ContainerTypeRepetitiveTask
    - ContainerTypeLongTask
    - ContainerTypeDirection
    - ContainerTypeState
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Task management dashboard API server.
    **Authentication:** Use `POST /users/login`, copy `accessToken` from the response, then click **Authorize** and paste it as the `access_token` cookie value. Alternatively, login from Swagger on the same host and cookies are sent automatically.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Dashboard API
  version: "1.0"
paths:
...
  /add-anonymous-task:
    put:
      consumes:
      - application/json
      description: Creates a simple anonymous task with default values
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Task'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create anonymous task
      tags:
      - tasks
...
  /change-tasks-order:
    post:
      consumes:
      - application/json
      description: Changes the order of tasks in a container
      parameters:
      - description: Change tasks order request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangeTasksOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Change tasks order
      tags:
      - tasks
...
  /done-tasks:
    get:
      consumes:
      - application/json
      description: Returns all done tasks where doneDateTime is today
      parameters:
      - description: From date-time filter (RFC3339)
        example: "2026-04-23T00:00:00+03:00"
        format: date-time
        in: query
        name: from
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Task'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get done tasks from today
      tags:
      - tasks
...
  /finish-task/{id}:
    put:
      consumes:
      - application/json
      description: Marks a task and all its descendants as done
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TaskResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Finish task recursively
      tags:
      - tasks
  /finish-tasks-by-ids/:
    put:
      consumes:
      - application/json
      description: Marks tasks by their IDs and all their descendants as done
      parameters:
      - description: List of task IDs
        in: body
        name: request
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Finish tasks by IDs recursively
      tags:
      - tasks
  /finish-tasks/:
    put:
      consumes:
      - application/json
      description: Marks multiple tasks and all their descendants as done
      parameters:
      - description: List of task IDs
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/handlers.TaskIDRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Finish multiple tasks recursively
      tags:
      - tasks
...
  /get-tasks:
    post:
      consumes:
      - application/json
      description: Returns multiple tasks by their IDs
      parameters:
      - description: List of task IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.TaskResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get tasks by IDs
      tags:
      - tasks
...
  /new-hierarchical-tasks:
    post:
      consumes:
      - application/json
      description: Creates a tree of tasks under a parent container
      parameters:
      - description: Hierarchical tasks creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.NewHierarchicalTasksRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TaskFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AccessTokenCookie: []
      summary: Create hierarchical tasks
      tags:
      - tasks
...
  /new-task:
    post:
      consumes:
      - application/json
      description: Creates a new task with optional parent relationship
      parameters:
      - description: Task creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.NewTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create new task
      tags:
      - tasks
...
  /task-report/{id}:
    get:
      consumes:
      - application/json
      description: Returns a tree of done descendant tasks under the root task. The
        response body is JSON null when there are no done descendants.
      parameters:
      - description: Root task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TaskReportTreeNode'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Task completion report tree
      tags:
      - tasks
  /task/{id}:
    get:
      consumes:
      - application/json
      description: Returns a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TaskFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get task by ID
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: Partially updates a task's description and/or notes
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PatchTaskByIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TaskFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Patch task by ID
      tags:
      - tasks
  /tasks/by-due-date:
    get:
      consumes:
      - application/json
      description: Returns open (not done) tasks whose due date falls on the specified
        calendar day (UTC)
      parameters:
      - description: Due date (YYYY-MM-DD, UTC)
        example: "2026-07-15"
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TaskFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AccessTokenCookie: []
      summary: Get open tasks by due date
      tags:
      - tasks
...
  /update-task:
    put:
      consumes:
      - application/json
      description: Updates an existing task by ID
      parameters:
      - description: Task update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.TaskPartial'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TaskFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update task
      tags:
      - tasks
...
schemes:
- http
- https
security:
- AccessTokenCookie: []
securityDefinitions:
  AccessTokenCookie:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"